Final Draft for Now
was able to get most of the functionality to work however there are
some bug, board knowns there is a winner because I can console if the
game is over or not and get the right results, however my winning
message does not populate. The play again function is not properly
running when the button is click the game reset in an incorrect way and
does not allow you to play again well. finally my save game(update)
function doesn't seem to work well either. other than that everything
else is fine, did some light css to make the game look decent.