Possible Final Draft
Was able to fix all the bugs that were present at the moment of
submitting the project. Awaiting to be tested by instructor before
deploying the final copy.

Fixed the clicking issue that occured after winning condition or draw,
changed a lot what shows and hides and when to meet the requirements.
Fixed giving the user feed back after win/loss or turn or anything
related to the game. Also fixed the console log bug that was there.